Removed noindex Flag from A/B Tests + Removed Google Optimize (#294)
* LPE-3963 Removed noindex flag for A/B tests

Since we’re using canonical tags with A/B test variants and marketing is only publishing variants that they’re actively using, and variants from concluded experiments redirect to the winner, there’s no need for the noindex flag in that context.

This was causing an issue with /home-page-v2 displaying the noindex flag on our homepage.

* Removed Google Optimize support

Google Optimize was sunset in 2023. We replaced this with a custom A/B test solution. No use keeping the code around.